Javadocs added + made constructors for some of the classes
I added javadocs to all of the model's, controller's and the ResourceLoader classes. Additionally i made constructors for several building classes. Also changed the return type from int to void of some methods that change the budget of the player, because the budget is an Integer type which is a reference.